Separate out Http into its own module
