+++ azure-functions-sdk/src/commands/new.rs
@@ -17,7 +17,10 @@ use std::{
 use syn::{parse::Parser, parse_file, punctuated::Punctuated, Ident, Item, Token};
 
 mod blob;
+mod http;
+
 pub use self::blob::Blob;
+pub use self::http::Http;
 
 fn get_path_for_function(name: &str) -> Result<String, String> {
     if !Regex::new("^[a-zA-Z][a-zA-Z0-9_]*$")
@@ -210,59 +213,7 @@ impl<'a> From<&'a ArgMatches<'a>> for New<'a> {
     }
 }
 
-struct Http<'a> {
-    name: &'a str,
-    auth_level: &'a str,
-}
 
-impl<'a> Http<'a> {
-    pub fn create_subcommand<'b>() -> App<'a, 'b> {
-        SubCommand::with_name("http")
-            .about("Creates a new HTTP triggered Azure Function.")
-            .arg(
-                Arg::with_name("name")
-                    .long("name")
-                    .short("n")
-                    .value_name("NAME")
-                    .help("The name of the new Azure Function.")
-                    .required(true),
-            )
-            .arg(
-                Arg::with_name("auth-level")
-                    .long("auth-level")
-                    .value_name("LEVEL")
-                    .possible_values(&["anonymous", "function", "admin"])
-                    .help("The authentication level for the HTTP function. Default is 'function'."),
-            )
-    }
-
-    pub fn execute(&self, quiet: bool) -> Result<(), String> {
-        let data = json!({
-            "name": self.name,
-            "auth_level": self.auth_level
-        });
-
-        create_function(self.name, "http.rs", &data, quiet)
-    }
-}
-
-impl<'a> From<&'a ArgMatches<'a>> for Http<'a> {
-    fn from(args: &'a ArgMatches<'a>) -> Self {
-        Http {
-            name: args.value_of("name").unwrap(),
-            auth_level: match args.value_of("auth-level") {
-                Some(level) => {
-                    if level == "function" {
-                        ""
-                    } else {
-                        level
-                    }
-                }
-                None => "",
-            },
-        }
-    }
-}
 
 struct Queue<'a> {
     name: &'a str,
